Migrate Text, Node and View Styles (#1519)
Fixes Migrate StyledTextRun to dc_bundle #1460
Towards Migrate ViewStyle/View to dc_bundle #1456
