+++ crates/dc_bundle/src/legacy_definition/view/node_style.rs
@@ -14,6 +14,22 @@
  * limitations under the License.
  */
 
+use crate::definition::element::{FontFeature, FontStyle, Hyperlink, TextDecoration};
+use crate::definition::layout::FlexWrap;
+use crate::legacy_definition::element::background::Background;
+use crate::legacy_definition::element::font::{FontStretch, FontWeight};
+use crate::legacy_definition::element::geometry::Size;
+use crate::legacy_definition::element::path::{LineHeight, Stroke};
+use crate::legacy_definition::element::variable::NumOrVar;
+use crate::legacy_definition::interaction::pointer::PointerEvents;
+use crate::legacy_definition::layout::grid::{GridLayoutType, GridSpan};
+use crate::legacy_definition::layout::positioning::{LayoutSizing, Overflow};
+use crate::legacy_definition::modifier::blend::BlendMode;
+use crate::legacy_definition::modifier::filter::FilterOp;
+use crate::legacy_definition::modifier::shadow::{BoxShadow, TextShadow};
+use crate::legacy_definition::modifier::text::{TextAlign, TextAlignVertical, TextOverflow};
+use crate::legacy_definition::modifier::transform::LayoutTransform;
+use crate::legacy_definition::plugin::meter_data::MeterData;
 use serde::{Deserialize, Serialize};
 
 #[derive(Copy, Clone, PartialEq, Debug, Serialize, Deserialize)]
@@ -39,5 +55,103 @@ pub enum Display {
 impl Default for Display {
     fn default() -> Self {
         Self::Flex
+    }
+}
+
+/// ToolkitStyle contains all of the styleable parameters accepted by the Rect and Text components.
+///
+#[derive(Clone, Debug, PartialEq, Deserialize, Serialize)]
+pub struct NodeStyle {
+    pub text_color: Background,
+    pub font_size: NumOrVar,
+    pub font_family: Option<String>,
+    pub font_weight: FontWeight,
+    pub font_style: FontStyle,
+    pub text_decoration: TextDecoration,
+    pub letter_spacing: Option<f32>,
+    pub font_stretch: FontStretch,
+    pub background: Vec<Background>,
+    pub box_shadow: Vec<BoxShadow>,
+    pub stroke: Stroke,
+    pub opacity: Option<f32>,
+    pub transform: Option<LayoutTransform>,
+    pub relative_transform: Option<LayoutTransform>,
+    pub text_align: TextAlign,
+    pub text_align_vertical: TextAlignVertical,
+    pub text_overflow: TextOverflow,
+    pub text_shadow: Option<TextShadow>,
+    pub node_size: Size<f32>,
+    pub line_height: LineHeight,
+    pub line_count: Option<usize>, // None means no limit on # lines.
+    pub font_features: Vec<FontFeature>,
+    pub filter: Vec<FilterOp>,
+    pub backdrop_filter: Vec<FilterOp>,
+    pub blend_mode: BlendMode,
+    pub hyperlink: Option<Hyperlink>,
+
+    pub display_type: Display,
+    pub flex_wrap: FlexWrap,
+    pub grid_layout: Option<GridLayoutType>,
+    pub grid_columns_rows: u32,
+    pub grid_adaptive_min_size: u32,
+    pub grid_span_content: Vec<GridSpan>,
+    pub overflow: Overflow,
+    pub max_children: Option<u32>,
+    pub overflow_node_id: Option<String>,
+    pub overflow_node_name: Option<String>,
+    pub cross_axis_item_spacing: f32,
+    pub horizontal_sizing: LayoutSizing,
+    pub vertical_sizing: LayoutSizing,
+    pub aspect_ratio: Number,
+    pub pointer_events: PointerEvents,
+    pub meter_data: Option<MeterData>,
+}
+
+impl Default for NodeStyle {
+    fn default() -> NodeStyle {
+        NodeStyle {
+            text_color: Background::None,
+            font_size: NumOrVar::Num(18.0),
+            font_family: None,
+            font_weight: FontWeight::NORMAL,
+            font_style: FontStyle::Normal,
+            text_decoration: TextDecoration::None,
+            letter_spacing: None,
+            font_stretch: FontStretch::NORMAL,
+            background: Vec::new(),
+            box_shadow: Vec::new(),
+            stroke: Stroke::default(),
+            opacity: None,
+            transform: None,
+            relative_transform: None,
+            text_align: TextAlign::Left,
+            text_align_vertical: TextAlignVertical::Top,
+            text_overflow: TextOverflow::Clip,
+            text_shadow: None,
+            node_size: Size::default(),
+            line_height: LineHeight::Percent(1.0),
+            line_count: None,
+            font_features: Vec::new(),
+            filter: Vec::new(),
+            backdrop_filter: Vec::new(),
+            blend_mode: BlendMode::default(),
+            hyperlink: None,
+            display_type: Display::default(),
+            flex_wrap: FlexWrap::NoWrap,
+            grid_layout: None,
+            grid_columns_rows: 0,
+            grid_adaptive_min_size: 1,
+            grid_span_content: vec![],
+            overflow: Overflow::default(),
+            max_children: None,
+            overflow_node_id: None,
+            overflow_node_name: None,
+            cross_axis_item_spacing: 0.0,
+            horizontal_sizing: LayoutSizing::default(),
+            vertical_sizing: LayoutSizing::default(),
+            aspect_ratio: Number::default(),
+            pointer_events: PointerEvents::default(),
+            meter_data: None,
+        }
     }
 }
